# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Todo {
  id: number;
  title: string;
  completed: boolean;
  createdAt: string;
}

export interface CreateTodoBody {
  /**
   * @minLength 1
   * @maxLength 500
   */
  title: string;
}

export interface UpdateTodoBody {
  /**
   * @minLength 1
   * @maxLength 500
   */
  title?: string;
  completed?: boolean;
}

export interface TodoSummary {
  total: number;
  active: number;
  completed: number;
}

export type ListTodosParams = {
  status?: ListTodosStatus;
};

export type ListTodosStatus =
  (typeof ListTodosStatus)[keyof typeof ListTodosStatus];

export const ListTodosStatus = {
  all: "all",
  active: "active",
  completed: "completed",
} as const;