# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: todos
    description: Todo operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /todos:
    get:
      operationId: listTodos
      tags: [todos]
      summary: List all todos
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [all, active, completed]
      responses:
        "200":
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
    post:
      operationId: createTodo
      tags: [todos]
      summary: Create a new todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTodoBody"
      responses:
        "201":
          description: Created todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
  /todos/{id}:
    patch:
      operationId: updateTodo
      tags: [todos]
      summary: Update a todo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTodoBody"
      responses:
        "200":
          description: Updated todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
    delete:
      operationId: deleteTodo
      tags: [todos]
      summary: Delete a todo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /todos/summary:
    get:
      operationId: getTodoSummary
      tags: [todos]
      summary: Get summary counts
      responses:
        "200":
          description: Summary of todos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoSummary"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Todo:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        completed:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - completed
        - createdAt
    CreateTodoBody:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
      required:
        - title
    UpdateTodoBody:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        completed:
          type: boolean
    TodoSummary:
      type: object
      properties:
        total:
          type: integer
        active:
          type: integer
        completed:
          type: integer
      required:
        - total
        - active
        - completed