# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all todos
 */
export const ListTodosQueryParams = zod.object({
  status: zod.enum(["all", "active", "completed"]).optional(),
});

export const ListTodosResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  completed: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListTodosResponse = zod.array(ListTodosResponseItem);

/**
 * @summary Create a new todo
 */
export const createTodoBodyTitleMax = 500;

export const CreateTodoBody = zod.object({
  title: zod.string().min(1).max(createTodoBodyTitleMax),
});

/**
 * @summary Update a todo
 */
export const UpdateTodoParams = zod.object({
  id: zod.coerce.number(),
});

export const updateTodoBodyTitleMax = 500;

export const UpdateTodoBody = zod.object({
  title: zod.string().min(1).max(updateTodoBodyTitleMax).optional(),
  completed: zod.boolean().optional(),
});

export const UpdateTodoResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  completed: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a todo
 */
export const DeleteTodoParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get summary counts
 */
export const GetTodoSummaryResponse = zod.object({
  total: zod.number(),
  active: zod.number(),
  completed: zod.number(),
});