# lib/api-zod/src/generated/types/todo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Todo {
  id: number;
  title: string;
  completed: boolean;
  createdAt: Date;
}