# lib/api-zod/src/generated/types/updateTodoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTodoBody {
  /**
   * @minLength 1
   * @maxLength 500
   */
  title?: string;
  completed?: boolean;
}